💩feat: 3-1 과제 제출

+++ src/main/kotlin/com/teamsparta/spatodolist/domain/todo/reply/service/ReplyService.kt
@@ -1,6 +1,7 @@
 package com.teamsparta.spatodolist.domain.todo.reply.service
 
 import com.teamsparta.courseregistration.domain.exception.ModelNotFoundException
+import com.teamsparta.spatodolist.domain.security.UserPrincipal
 import com.teamsparta.spatodolist.domain.todo.reply.dtos.CreateReplyArgument
 import com.teamsparta.spatodolist.domain.todo.reply.dtos.DeleteReplyArgument
 import com.teamsparta.spatodolist.domain.todo.reply.dtos.ReplyDto
@@ -8,6 +9,7 @@ import com.teamsparta.spatodolist.domain.todo.reply.dtos.UpdateReplyArgument
 import com.teamsparta.spatodolist.domain.todo.reply.model.Reply
 import com.teamsparta.spatodolist.domain.todo.reply.repository.ReplyRepository
 import com.teamsparta.spatodolist.domain.todo.todocard.repository.TodoCardRepository
+import com.teamsparta.spatodolist.domain.users.repository.UserRepository
 import jakarta.transaction.Transactional
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
@@ -16,43 +18,56 @@ import org.springframework.stereotype.Service
 class ReplyService(
     private val replyRepository: ReplyRepository,
     private val todoRepository: TodoCardRepository,
+    private val userRepository: UserRepository,
 ) {
 
     @Transactional
-    fun createReply(todoCardId: Long, argument: CreateReplyArgument): ReplyDto {
+    fun createReply(todoCardId: Long, argument: CreateReplyArgument, userPrincipal: UserPrincipal): ReplyDto {
         val foundTodoCard =
             todoRepository.findByIdOrNull(todoCardId) ?: throw ModelNotFoundException("TodoCard", todoCardId)
+        val users =
+            userRepository.findByIdOrNull(userPrincipal.id) ?: throw ModelNotFoundException("User", userPrincipal.id)
+
 
         val reply = Reply(
             argument.content,
             argument.authorName,
             argument.password,
-            foundTodoCard
+            foundTodoCard,
+            users
         )
 
         val savedReply = replyRepository.save(reply)
-
         return ReplyDto.from(savedReply)
     }
 
     @Transactional
-    fun updateReply(replyId: Long, todoCardId: Long, argument: UpdateReplyArgument): ReplyDto {
+    fun updateReply(
+        replyId: Long,
+        todoCardId: Long,
+        argument: UpdateReplyArgument,
+        userPrincipal: UserPrincipal,
+    ): ReplyDto {
         val foundTodoCard =
             todoRepository.findByIdOrNull(todoCardId) ?: throw ModelNotFoundException("TodoCard", todoCardId)
         val foundReply = replyRepository.findByIdOrNull(replyId) ?: throw ModelNotFoundException("Reply", replyId)
 
-        foundReply.checkAuthentication(argument.authorName, argument.password)
+        foundReply.checkAuthorization(userPrincipal.to())
+
+        //foundReply.checkAuthorization(argument.authorName, argument.password)
         foundReply.changeContent(argument.content)
 
         return ReplyDto.from(foundReply)
     }
 
-    fun deleteReply(replyId: Long, todoCardId: Long, argument: DeleteReplyArgument) {
+    fun deleteReply(replyId: Long, todoCardId: Long, argument: DeleteReplyArgument, userPrincipal: UserPrincipal) {
         val foundTodoCard =
             todoRepository.findByIdOrNull(todoCardId) ?: throw ModelNotFoundException("TodoCard", todoCardId)
         val foundReply = replyRepository.findByIdOrNull(replyId) ?: throw ModelNotFoundException("Reply", replyId)
 
-        foundReply.checkAuthentication(argument.authorName, argument.password)
+        foundReply.checkAuthorization(userPrincipal.to())
+
+        //foundReply.checkAuthorization(argument.authorName, argument.password)
 
         replyRepository.deleteById(replyId)
     }
